requirements/base.txt
psycopg2==2.5.4
Pillow==2.5.3
Django==1.7.1
django_compressor==1.4
# Required by django-compressor
django-appconf==0.6
django-localflavor==1.0
six==1.6.1

BeautifulSoup4==4.3.2
